move set file
