+++ front-end/src/components/exercises/ExerciseCollapseChild.tsx
@@ -3,7 +3,7 @@ import React, { useEffect, useState } from "react";
 import { Button, Collapse } from "antd";
 import { useParams } from "react-router";
 import { TUsersExerciseData } from "../../api/types";
-import { Set } from "./sets/Set";
+import { Set } from "./Set";
 import { CheckCircleOutlined } from "@ant-design/icons";
 import { arrToNum } from "../../utils/utils";
